fscrypt-crypt-util: add HCTR2 tests
This patch is not intended to be accepted yet.  It is reliant on HCTR2
support in the kernel which has not yet been accepted.  See the HCTR2
patch here:
https://lore.kernel.org/all/20220412172816.917723-1-nhuck@google.com/

HCTR2 is a new wide-block encryption mode that can used for filename
encryption in fscrypt.  This patch adds tests for fscrypt policies using
HCTR2 for filename encryption.

More information on HCTR2 can be found here:
Length-preserving encryption with HCTR2
https://ia.cr/2021/1441